+++ demo.py
@@ -4,7 +4,7 @@ from mapmatching import build_geograph, ST_Matching
 # 方法1：
 # 根据 bbox 从 OSM 下载路网，从头解析获得路网数据
 # net = build_geograph(bbox=[113.930914, 22.570536, 113.945456, 22.585613],
-#                      xml_fn="./data/network/LXD.osm.xml", ll=False)
+#                      xml_fn="./data/network/LXD.osm.xml", ll=False, n_jobs=1)
 # 将预处理路网保存为 ckpt
 # net.save_checkpoint('./data/network/LXD_graph.ckpt')
 
@@ -23,7 +23,7 @@ res = matcher.matching(traj, top_k=5, dir_trans=True, details=False, plot=True,
 
 # 后续步骤可按需选择
 """step 4: 将轨迹点映射到匹配道路上"""
-path = matcher.transform_res_2_path(res, ori_crs=True)
+path = matcher.transform_res_2_path(res)
 proj_traj = matcher.project(traj, path)
 
 """step 5: eval"""